+++ packages/firebase-dart/example/auth/index.dart
@@ -28,6 +28,7 @@ class AuthApp {
   final AnchorElement logout;
   final TableElement authInfo;
   final ParagraphElement error;
+  final SelectElement persistenceState;
   final ButtonElement verifyEmail;
 
   AuthApp()
@@ -38,6 +39,7 @@ class AuthApp {
         this.error = querySelector("#register_form p"),
         this.logout = querySelector("#logout_btn"),
         this.registerForm = querySelector("#register_form"),
+        this.persistenceState = querySelector("#persistent_state"),
         this.verifyEmail = querySelector('#verify_email') {
     logout.onClick.listen((e) {
       e.preventDefault();
@@ -62,8 +64,20 @@ class AuthApp {
     verifyEmail.onClick.listen((e) async {
       verifyEmail.disabled = true;
       verifyEmail.text = 'Sending verification email...';
-      await auth.currentUser.sendEmailVerification();
-      verifyEmail.text = 'Verification email sent!';
+      try {
+        // you will get the verification email in czech language
+        auth.languageCode = 'cs';
+        // url should be any authorized domain in your console - we use here,
+        // for this example, authDomain because it is whitelisted by default
+        // More info: https://firebase.google.com/docs/auth/web/passing-state-in-email-actions
+        await auth.currentUser.sendEmailVerification(
+            new fb.ActionCodeSettings(url: "https://$authDomain"));
+        verifyEmail.text = 'Verification email sent!';
+      } catch (e) {
+        verifyEmail
+          ..text = e.toString()
+          ..style.color = 'red';
+      }
     });
   }
 
@@ -71,6 +85,10 @@ class AuthApp {
     if (email.isNotEmpty && password.isNotEmpty) {
       var trySignin = false;
       try {
+        // Modifies persistence state. More info: https://firebase.google.com/docs/auth/web/auth-state-persistence
+        var selectedPersistence =
+            persistenceState.options[persistenceState.selectedIndex].value;
+        await auth.setPersistence(selectedPersistence);
         await auth.createUserWithEmailAndPassword(email, password);
       } on fb.FirebaseError catch (e) {
         if (e.code == "auth/email-already-in-use") {